Prevent rendering of request buttons for SPEC titles with no items
This addresses a situation that can arise from dubious data that
was reported by a tester in searchworks-stage for instance 14642011.

We normally allow a request to be made if we know an item is on
order but it hasn't been added to FOLIO yet, but SPEC is the only
library where this policy should not apply.

Closes #4205